Fixed major bug, perfected the pathfind steering

+++ MemeLib/MemeLib-Game/GameApp.h
@@ -36,7 +36,7 @@ private:
 	NavMesh*	mNavMesh;
 	Path		m_path;
 
-	size_t	m_index = 0;
+	int	m_index = 0;
 	Node*	mp_target;
 };
